removed unnecessary casting.

+++ src/algorithms/imageProcessing/CurvatureScaleSpaceCornerDetector.java
@@ -199,8 +199,8 @@ public class CurvatureScaleSpaceCornerDetector extends
                 int idx = subList.get(j).intValue();
                 // exclude if k < 0.1?
                 //if (Math.abs(scaleSpace.getK(idx)) > 0.1) {
-                    int xte = (int) Math.round(scaleSpace.getX(idx));
-                    int yte = (int) Math.round(scaleSpace.getY(idx));            
+                    int xte = Math.round(scaleSpace.getX(idx));
+                    int yte = Math.round(scaleSpace.getY(idx));            
                     cornersForMatching.add(xte, yte);
                 //}
             }
@@ -272,8 +272,8 @@ public class CurvatureScaleSpaceCornerDetector extends
                                  
         //store xc and yc for the edge
         for (int ii = 0; ii < candidateCornersXY.getN(); ii++) {
-            int xte = (int) Math.round(candidateCornersXY.getX(ii));
-            int yte = (int) Math.round(candidateCornersXY.getY(ii)); 
+            int xte = Math.round(candidateCornersXY.getX(ii));
+            int yte = Math.round(candidateCornersXY.getY(ii)); 
             edgeCorners.add(xte, yte);
         }
